requirements.txt
pycryptodome>=3.20.0
kybercffi>=0.3.0